refactored grips as well
